search: Limit concurrency before spinning up a goroutine
This will prevent us starting a large number of goroutine which wait. It will
also limit concurrency before we hit gitserver. Additionally the limit is not
hardcoded, but rather is a multiple of the number of search replicas. However,
this does remove our global rate limiter. The global rate limiter was more an
artifact of implementation, than an intended implementation. So a per request
rate limiter should be sufficient.